requirements.txt
click==8.0.3
cycler==0.10.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
PySastrawi==1.2.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.8
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.0.0
tqdm==4.62.3
wordcloud==1.8.1
